+++ ynab_commands/config.py
@@ -1,5 +1,10 @@
+import logging
+from pathlib import Path
+
 from pydantic.env_settings import BaseSettings
 from pydantic.types import SecretStr
+
+log = logging.getLogger(__name__)
 
 
 class Config(BaseSettings):
@@ -12,3 +17,6 @@ class Config(BaseSettings):
     splitwise_consumer_key: SecretStr = SecretStr("123")
     splitwise_consumer_secret: SecretStr = SecretStr("123")
     splitwise_api_key: SecretStr = SecretStr("123")
+
+
+CONFIG = Config(_env_file=Path(__file__).parents[1] / "prod.env")  # type: ignore[call-arg]