+++ ynab_commands/ynab_api.py
@@ -1,17 +1,24 @@
 import json
+import logging
 from typing import Any
 
 import requests
 import requests_cache
 from pydantic.types import SecretStr
 
+from ynab_commands.config import CONFIG
 from ynab_commands.models import (
     Account,
     BudgetSummaryResponse,
+    PatchTransactionWrapper,
+    SaveSubTransaction,
+    SaveTransactionWithIdOrImportId,
     SaveTransactionWrapper,
     TransactionDetail,
     TransactionsResponse,
 )
+
+log = logging.getLogger(__name__)
 
 
 def parse_payload(**kwargs: Any) -> dict[str, Any]:
@@ -32,7 +39,7 @@ def get(
         headers={"Authorization": f"Bearer {token.get_secret_value()}"},
     )
     response.raise_for_status()
-
+    log.debug(f"GET response: {response.status_code}.")
     if response.status_code != 200:
         raise requests.RequestException
 
@@ -51,6 +58,27 @@ def put(
         "Content-Type": "application/json",
     }
     response = session.put(url=url, headers=headers, data=data)
+    log.debug(f"PUT response: {response.status_code}.")
+    response.raise_for_status()
+    if response.status_code != 200:
+        raise requests.RequestException
+
+    return response.json()
+
+
+def patch(
+    data: Any,
+    session: requests.Session,
+    url: str,
+    token: SecretStr,
+) -> Any:
+    headers = {
+        "accept": "application/json",
+        "Authorization": f"Bearer {token.get_secret_value()}",
+        "Content-Type": "application/json",
+    }
+    response = session.patch(url=url, headers=headers, data=data)
+    log.debug(f"PATCH response: {response.status_code}.")
     response.raise_for_status()
     if response.status_code != 200:
         raise requests.RequestException
@@ -108,6 +136,7 @@ class YNABApi:
         """
         url = f"{self._base_url}/budgets/{budget_id}/transactions"
         payload = parse_payload(**kwargs)
+        log.debug("Grabbing budget transactions.")
 
         response_json = get(
             session=self._session, url=url, token=self._token, params=payload
@@ -115,16 +144,67 @@ class YNABApi:
 
         return TransactionsResponse(**response_json["data"])
 
-    def update_transaction(
+    def _update_transactions(
         self,
         budget_id: str,
-        transaction_id: str,
-        updated_transaction: SaveTransactionWrapper,
-    ) -> TransactionDetail:
-        url = f"{self._base_url}/budgets/{budget_id}/transactions/{transaction_id}"
-        data = parse_transaction(updated_transaction)
-        response_json = put(
-            session=self._session, url=url, token=self._token, data=data
+        transactions: list[SaveTransactionWithIdOrImportId],
+    ) -> Any:
+        url = f"{self._base_url}/budgets/{budget_id}/transactions"
+        transactions_obj = PatchTransactionWrapper(transactions=transactions).dict()
+        log.debug("Patching transactions")
+        response_json = patch(
+            session=self._session,
+            url=url,
+            token=self._token,
+            data=json.dumps(transactions_obj),
         )
 
-        return TransactionDetail(**response_json["data"]["transaction"])
+        return response_json
+
+    def split_and_update_transaction(
+        self,
+        filtered_response: TransactionsResponse,
+    ) -> None:
+        total_in_pounds = repr(filtered_response.transaction_total)
+        updated_transactions = [self._split_transaction(t) for t in filtered_response]
+        log.debug(
+            "%s transactions worth %s about to be updated.",
+            len(updated_transactions),
+            total_in_pounds,
+        )
+
+        self._update_transactions(
+            budget_id=CONFIG.budget_id,
+            transactions=updated_transactions,
+        )
+
+        print(f"Processed {len(filtered_response)} transactions")
+        print(f"Add {total_in_pounds} to splitwise")
+
+    def _split_transaction(
+        self, transaction: TransactionDetail
+    ) -> SaveTransactionWithIdOrImportId:
+        split_amount = transaction.amount // 2
+        personal_subtransaction = SaveSubTransaction(
+            amount=split_amount,
+            category_id=transaction.category_id,
+            payee_id=transaction.payee_id,
+            payee_name=transaction.payee_name,
+            memo=transaction.memo,
+        )
+        splitwise_subtransaction = SaveSubTransaction(
+            amount=split_amount,
+            category_id=CONFIG.splitwise_id,
+            payee_id=transaction.payee_id,
+            payee_name=transaction.payee_name,
+            memo="Auto-split",
+        )
+        return SaveTransactionWithIdOrImportId(
+            id=transaction.id,
+            amount=transaction.amount,
+            account_id=transaction.account_id,
+            date=transaction.date,
+            approved=True,
+            cleared="cleared",
+            subtransactions=[personal_subtransaction, splitwise_subtransaction],
+        )