+++ changelog.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.2]
+### Added
+- Add MilliUnits object
+- logging debug messages to stdout
+- Patch updating all transactions to splitwise
+
+### Changed
+- moved splitting logic to ynab api class
+
 ## [0.2.1]
 ### Added
 - add printing total amount